+++ containers/gesture-recognition/models/anchors_palm.npy
