+++ containers/gesture-recognition/models/mediapipe_hand_gesture_float_CannedGestureClassifier.tflite
