+++ containers/gesture-recognition/models/mediapipe_hand_gesture_w8a8_PalmDetector.tflite
