+++ containers/gesture-recognition/requirements.txt
