+++ containers/gesture-recognition/utils/constants.py
