+++ src/arduino/app_bricks/gesture_recognition/README.md
@@ -1,0 +1,5 @@
+# Gesture Recognition Brick
+
+This gesture recognition brick utilizes a pre-trained model to analyze video streams from a camera.
+The output is a video stream featuring gesture recognition as overlay, with the added capability to
+trigger actions based on these detections.
