+++ src/arduino/app_bricks/gesture_recognition/brick_config.yaml
@@ -1,0 +1,10 @@
+id: arduino:gesture_recognition
+name: Gesture Recognition
+description: |
+  This gesture recognition brick utilizes a pre-trained model to analyze video streams from a camera.
+  The output is a video stream featuring gesture recognition as overlay, with the added capability to
+  trigger actions based on these detections.
+category: video
+requires_container: true
+required_devices:
+  - camera
