+++ src/arduino/app_bricks/gesture_recognition/examples/02_enter_exit.py
@@ -2,13 +2,13 @@
 #
 # SPDX-License-Identifier: MPL-2.0
 
-# EXAMPLE_NAME = "Hand enter/exit detection"
+# EXAMPLE_NAME = "Gesture recognition hand enter/exit detection"
 # EXAMPLE_REQUIRES = "Requires a connected camera"
 
-from arduino.app_bricks.hand_gesture_tracking import HandGestureTracking
+from arduino.app_bricks.gesture_recognition import GestureRecognition
 from arduino.app_utils.app import App
 
-pd = HandGestureTracking()
+pd = GestureRecognition()
 pd.on_enter(lambda: print("Hi there!"))
 pd.on_exit(lambda: print("Goodbye!"))
