+++ src/arduino/app_bricks/gesture_recognition/examples/03_confirm_deny.py
@@ -5,10 +5,10 @@
 # EXAMPLE_NAME = "Get user confirmation with hand gestures"
 # EXAMPLE_REQUIRES = "Requires a connected camera"
 
-from arduino.app_bricks.hand_gesture_tracking import HandGestureTracking
+from arduino.app_bricks.gesture_recognition import GestureRecognition
 from arduino.app_utils.app import App
 
-pd = HandGestureTracking()
+pd = GestureRecognition()
 pd.on_gesture("Thumb_Up", lambda meta: print("Operation confirmed!"))
 pd.on_gesture("Thumb_Down", lambda meta: print("Operation denied!"))
