--- src/arduino/app_bricks/hand_gesture_tracking/README.md
@@ -1,1 +1,0 @@
-# TODO
