--- src/arduino/app_bricks/hand_gesture_tracking/brick_config.yaml
@@ -1,10 +1,0 @@
-id: arduino:hand_gesture_tracking
-name: Hand gesture tracking
-description: |
-  This hand gesture tracking brick utilizes a pre-trained model to analyze video streams from a camera.
-  The output is a video stream featuring hand gesture tracking as overlay, with the added capability to
-  trigger actions based on these detections.
-category: video
-requires_container: true
-required_devices:
-  - camera
